Removed the main() from the ziggurat.c file as you then can't link against it since you end up with duplicate _main symbols. (doh)
Moved the test code to ziggurat_test.c and updated the Makefile to build the
whole thing.

+++ ziggurat.h
@@ -24,4 +24,4 @@
 
 void ziggurat_init( const long rand_seed );
 
-int32_t ziggurat_next();
+double ziggurat_next();
